HepMC -> ROOT conversion
Compiles and runs, but the resulting file is not recognized by
EntanglementNtupleProducer. More specifically, it does not seem to
recognize vector<reco::GenParticle> collection, which is labelled as
"genParticles".

HepMC format is what MadGraph produces out-of-the-box after Pythia8 has
run the showering and hadronization steps. I'm currently using version 2
of the format.