Get rid of the "Can not compute ppi(Q);" warning
The function that prompts this warning, PolarimetricA1::PVC(), call
BreitWigner method, but its output is never used to calculate the
returned value nor does the function modify the values of variables that
are paased to it because the function takes the arguments by value, not
by reference. Tested by producing a small sample of events. The results
did not change but the warning is now gone.

https://github.com/HEP-KBFI/tautau-Entanglement/issues/4