Run the analysis in bins of |cos(theta)|
Our initial plan was to run the measurement of each bin in a single
analysis job, which would effectively repeat the measurement for as many
times as there are bins in |cos(theta)|. However, it introduces numerous
problems because the most compute-intensive jobs are rho_rho, which take
approximately 19h to run and consume 1.5GB of memory. We might even fit
in the memory requirements of 2GB if we just store the
pointer/index/std::reference_wrapper<> per Data entry (albeit barely: 4
bytes x 12.5M events x ~7 duplication of events due to binning = 350 MB)
but the jobs would definitely run longer than 2 days. Another issue is
that we'd have to either extend the currnt JSON format to acommodate the
extra results from binned measurements, or create yet another output
file. Third, the implementation becomes quite convoluted and difficult
to debug in case we run into problems. The only solution that scales
well with the number of bins, doesn't require us to change the JSON
format and avoids excessive memory consumption and long runtime is to
create a new job for every cut in |cost(theta)|. The results can be
easily plotted by just parsing JSON files. In fact, this would allow us
to easily compare the significance vs |cos(theta)| cut and make more
informed decision when choosing the optimal cut. Later on, the analysis
can be rerun with just this one cut.
https://github.com/HEP-KBFI/tautau-Entanglement/issues/4